Add RecordViewModel date save test case
